Specify a rebar3 version in github action
This is to avoid "rebar3: command not found" in github actions.

Switch matrix representation so we can specify the rebar3 version,
os version and Erlang version independently of each other.

This will hopefully make it easier to set version: we get greater
flexibility, instead of having to find some minimum version that
works with all.