Shift workflow Erlang version from 23..25 to 24..26
In the github workflow actions, run on Erlang versions 24..26

From 26, Erlang now warns about unknown functions and types. Add the

  no_unknown

option to suppress those warnings, since otherwise, there will be
issues for functions and types from rebar3 itself, as these are
unknown when running dialyzer.

(I tried, but failed, to include rebar3 in the dialyzer run, so for
now, just disable those warnings.)

However, the no_unknown option is not a available in Erlang 25 and
earlier. Including it will cause an error that it is an unknown
warning. So turn rebar.config into rebar.config.script to be able
to conditionally include that option.